requirements.txt
# Core runtime (HTTP, env, YAML, validation)
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
validators==0.22.0
beautifulsoup4==4.12.2
lxml==5.2.1

# Media I/O and processing
Pillow==10.4.0
imageio==2.34.1
imageio-ffmpeg==0.4.9
numpy==1.26.4

# Data, progress, logging
pandas==2.2.2
tqdm==4.66.4
rich==13.7.1

# Audio pipeline
gTTS==2.5.1
pydub==0.25.1
mutagen==1.47.0

# Video assembly
moviepy==2.0.0
ffmpeg-python==0.2.0

# ----- Optional integrations (commented; enable as needed) -----

# Segmind Pixelflow (if using SDK rather than pure HTTP)
# segmind==0.2.0

# Amazon Product Advertising API
